Extract test from isk4
